project_6/bq_loader/requirements.txt
google-cloud-bigquery
cloudevents==1.11.0
functions_framework